+++ packages/lumx-react/src/components/notification/constants.ts
@@ -14,7 +14,7 @@ export const NOTIFICATION_CONFIGURATION = {
         icon: mdiAlert,
     },
     info: {
-        color: 'dark',
+        color: 'blue',
         icon: mdiInformation,
     },
     success: {