+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   `Lightbox`: fix a11y color contrast on the close button.
+
 ## [3.7.5][] - 2024-07-25
 
 ### Changed